Move メソッドをドメイン層に
